refactor: move otel file
